refactor: move Result of folder
